feat: finish adding dropdown for account details, refactor wallet main into several widgets, colocate related widgets for wallet main
